+++ api/src/setup/env.js
@@ -11,6 +11,6 @@ exports.EMAIL_PORT = process.env.EMAIL_PORT || 25;
 exports.EMAIL_USERNAME = process.env.EMAIL_USERNAME;
 exports.EMAIL_PASSWORD = process.env.EMAIL_PASSWORD;
 // CONFIG
-exports.STATIONSCHEDULE = process.env.STATIONCHEDULE || 1000 * 60 * 60 * 12;
-exports.AIRDATASCHEDULE = process.env.AIRDATASCHEDULE || 1000 * 60 * 30;
+exports.STATIONSCHEDULE = process.env.STATIONCHEDULE || 60 * 60 * 12;
+exports.AIRDATASCHEDULE = process.env.AIRDATASCHEDULE || 60 * 30;
 exports.TIMEZONE = process.env.TIMEZONE || 'Europe/Warsaw';
