+++ api/src/worker.js
@@ -1,8 +1,12 @@
 const db = require('./setup/db');
-const stationUpdater = require('./utils/jobs/stationUpdater');
+const env = require('./setup/env');
 const notifications = require('./utils/jobs/notifications');
+const stationLocationJob = require('./utils/jobs/stationLocation');
+const airIndexJob = require('./utils/jobs/stationAirIndex');
+const sensorDataJob = require('./utils/jobs/sensorData');
 
-// update station list
-stationUpdater.start();
-// notifications
 notifications.start();
+//TODO: te joby mają taką samą funkcję start(). Warto stworzyć jakiegoś jobSchedulera
+stationLocationJob.start(env.STATIONSCHEDULE, 15);
+airIndexJob.start(env.AIRDATASCHEDULE, 60);
+sensorDataJob.start(env.AIRDATASCHEDULE, 120);
